+++ src/pybdsim/Plot.py
@@ -481,7 +481,7 @@ def BDSIMOptics(rebdsimOpticsOutput, outputfilename=None, saveall=True, survey=N
         print("Written ", output_filename)
 
 
-def Histogram1D(histogram, xlabel=None, ylabel=None, title=None, scalingFactor=1.0, xScalingFactor=1.0, figsize=(6.4, 4.8), log=False, **errorbarKwargs):
+def Histogram1D(histogram, xlabel=None, ylabel=None, title=None, scalingFactor=1.0, xScalingFactor=1.0, figsize=(6.4, 4.8), log=False, ax=None, **errorbarKwargs):
     """
     Plot a pybdsim.Data.TH1 instance.
 
@@ -491,14 +491,19 @@ def Histogram1D(histogram, xlabel=None, ylabel=None, title=None, scalingFactor=1
     :param scalingFactor: multiplier for values
     :param xScalingFactor: multiplier for x axis coordinates
     :param log: whether to automatically plot on a vertical log scale
+    :param ax: Matplotlib.Axis instance to draw into. If None, a figure will be created.
 
     return figure instance
     """
     if 'drawstyle' not in errorbarKwargs:
         errorbarKwargs['drawstyle'] = 'steps-mid'
     h = histogram
-    f = _plt.figure(figsize=figsize)
-    ax = f.add_subplot(111)
+    
+    incomingAxis = bool(ax)
+    if not ax:
+        f = _plt.figure(figsize=figsize)
+        ax = f.add_subplot(111)
+    
     sf  = scalingFactor #shortcut
     xsf = xScalingFactor
     histEmpty = len(h.contents[h.contents!=0]) == 0
@@ -538,16 +543,18 @@ def Histogram1D(histogram, xlabel=None, ylabel=None, title=None, scalingFactor=1
         # with numerical precision something very small like 1e-22. Therefore,
         # just round down to next power of 10 on the contents.
         ymin = sf * 10 ** (_np.floor(_np.log10(_np.min(h.contents[h.contents > 0]))))
-        _plt.ylim(abs(ymin)*0.9,abs(ymax)*1.3)
+        ax.set_ylim(abs(ymin)*0.9,abs(ymax)*1.3)
         if _matplotlib.__version__ >= '3.3':
-            _plt.yscale('log', nonpositive='clip')
+            ax.set_yscale('log', nonpositive='clip')
         else:
-            _plt.yscale('log', nonposy='clip')
+            ax.set_yscale('log', nonposy='clip')
     else:
-        _plt.ylim(ymin*0.8, ymax*1.05)
+        ax.set_ylim(ymin*0.8, ymax*1.05)
 
-    _plt.tight_layout()
-    return f
+    if not incomingAxis:
+        _plt.tight_layout()
+    
+    return _plt.gcf()
 
 def Spectra(spectra, log=False, xlog=False, xlabel=None, ylabel=None, title=None,
             scalingFactors=None, xScalingFactors=None,
@@ -621,7 +628,9 @@ def Spectra(spectra, log=False, xlog=False, xlabel=None, ylabel=None, title=None
         return [f1]
 
 
-def Histogram1DMultiple(histograms, labels, log=False, xlog=False, xlabel=None, ylabel=None, title=None, scalingFactors=None, xScalingFactors=None, figsize=(10,5), legendKwargs={}, **errorbarKwargs):
+def Histogram1DMultiple(histograms, labels, log=False, xlog=False, xlabel=None, ylabel=None,
+                        title=None, scalingFactors=None, xScalingFactors=None, figsize=(10,5),
+                        legendKwargs={}, ax=None, **errorbarKwargs):
     """
     Plot multiple 1D histograms on the same plot. Histograms and labels should 
     be lists of the same length with pybdsim.Data.TH1 objects and strings.
@@ -639,12 +648,16 @@ def Histogram1DMultiple(histograms, labels, log=False, xlog=False, xlabel=None, 
                           scalingFactors=[1,100,100],
                           xScalingFactors=1e6,
                           log=True)
+
+    :param ax: matplotlib axes to draw into - if none, a new figure will be created.
     """
     if "xScalingFactor" in errorbarKwargs:
         raise ValueError("'xScalingFactor' - did you mean 'xScalingFactors'?")
 
-    f = _plt.figure(figsize=figsize)
-    ax = f.add_subplot(111)
+    incomingAxis = bool(ax)
+    if not ax:
+        f = _plt.figure(figsize=figsize)
+        ax = f.add_subplot(111)
     
     if scalingFactors is None:
         scalingFactors = _np.ones_like(histograms)
@@ -693,27 +706,33 @@ def Histogram1DMultiple(histograms, labels, log=False, xlog=False, xlabel=None, 
     else:
         ax.set_title(title)
     if log and not allHistsEmpty:
-        _plt.ylim(abs(yminpos)*0.9,abs(ymax)*1.1)
+        aymin = 10**_np.floor(_np.log10(abs(yminpos)*0.9))
+        aymax = 10**_np.ceil(_np.log10(abs(ymax)*1.1))
+        ax.set_ylim(aymin, aymax)
         if _matplotlib.__version__ >= '3.3':
-            _plt.yscale('log', nonpositive='clip')
+            ax.set_yscale('log', nonpositive='clip')
         else:
-            _plt.yscale('log', nonposy='clip')
+            ax.set_yscale('log', nonposy='clip')
     else:
-        _plt.ylim(ymin*0.8, ymax*1.05)
+        ax.set_ylim(ymin*0.8, ymax*1.05)
 
     if xlog:
-        _plt.xscale('log')
+        ax.set_xscale('log')
         suggestedXMin = 0.95*xmin
         if suggestedXMin <= 0:
             suggestedXMin = 1
-        _plt.xlim(suggestedXMin, 1.05*xmax)
+        ax.set_xlim(suggestedXMin, 1.05*xmax)
 
-    _plt.legend(**legendKwargs)
-    _plt.tight_layout()
+    ax.legend(**legendKwargs)
+    if not incomingAxis:
+        _plt.tight_layout()
     
-    return f
+    return _plt.gcf()
 
-def Histogram2D(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlabel="", ylabel="", zlabel="", title="", aspect="auto", scalingFactor=1.0, xScalingFactor=1.0, yScalingFactor=1.0, figsize=(6,5), vmin=None, autovmin=False, vmax=None, colourbar=True, **imshowKwargs):
+def Histogram2D(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlabel="", ylabel="",
+                zlabel="", title="", aspect="auto", scalingFactor=1.0, xScalingFactor=1.0,
+                yScalingFactor=1.0, figsize=(6,5), vmin=None, autovmin=False, vmax=None,
+                colourbar=True, ax=None, cax=None, **imshowKwargs):
     """
     Plot a pybdsim.Data.TH2 instance.
     logNorm        - logarithmic colour scale
@@ -727,11 +746,16 @@ def Histogram2D(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlab
     autovmin       - automatically determin the lower limit of the colourbar from the data
     vmin           - explicitly control the vmin for the colour normalisation
     vmax           - explicitly control the vmax for the colour normalisation
+    ax             - optional matplotlib axes to draw into
+    cax            - optional axes to draw coloubar into
 
     return figure instance
     """
     h = histogram
-    f = _plt.figure(figsize=figsize)
+    incomingAxis = bool(ax)
+    if not ax:
+        f = _plt.figure(figsize=figsize)
+        ax = f.add_subplot(111)
     x, y = _np.meshgrid(h.xcentres,h.ycentres)
     sf  = scalingFactor #shortcut
     xsf = xScalingFactor
@@ -751,15 +775,16 @@ def Histogram2D(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlab
     if logNorm:
         d = _copy.deepcopy(sf*h.contents.T)
         norm = _LogNorm(vmin=vmin,vmax=vmax) if vmax is not None else _LogNorm(vmin=vmin)
-        ax = f.add_subplot(111)
         im = ax.pcolormesh(h.xedges*xsf, h.yedges*ysf, d, norm=norm, rasterized=True, **imshowKwargs)
         #_plt.imshow(d, extent=ext, origin='lower', aspect=aspect, norm=norm, interpolation='none', **imshowKwargs)
         if colourbar:
-            _plt.colorbar(im, label=zlabel)
+            _plt.colorbar(im, label=zlabel, cax=cax)
     else:
-        _plt.imshow(sf*h.contents.T, extent=ext, origin='lower', aspect=aspect, interpolation='none', vmin=vmin, vmax=vmax,**imshowKwargs)
+        ax.imshow(sf*h.contents.T, extent=ext, origin='lower', aspect=aspect, interpolation='none', vmin=vmin, vmax=vmax,**imshowKwargs)
         if colourbar:
-            _plt.colorbar(format='%.0e', label=zlabel)
+            _plt.colorbar(format='%.0e', label=zlabel, cax=cax)
+
+    ax.set_aspect(aspect)
 
     if xLogScale:
         _plt.xscale('log')
@@ -767,27 +792,32 @@ def Histogram2D(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlab
         _plt.yscale('log')
 
     if xlabel == "":
-        _plt.xlabel(h.xlabel)
+        ax.set_xlabel(h.xlabel)
     elif xlabel is None:
         pass
     else:
-        _plt.xlabel(xlabel)
+        ax.set_xlabel(xlabel)
     if ylabel == "":
-        _plt.ylabel(h.ylabel)
+        ax.set_ylabel(h.ylabel)
     elif ylabel is None:
         pass
     else:
-        _plt.ylabel(ylabel)
+        ax.set_ylabel(ylabel)
     if title == "":
-        _plt.title(h.title) # default to one in histogram
+        ax.set_title(h.title) # default to one in histogram
     elif title is None:
         pass
     else:
-        _plt.title(title)
-    _plt.tight_layout()
-    return f
+        ax.set_title(title)
 
-def Histogram2DErrors(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlabel="", ylabel="", zlabel="", title="", aspect="auto", scalingFactor=1.0, xScalingFactor=1.0, yScalingFactor=1.0, figsize=(6,5), vmin=None, autovmin=False, vmax=None, **imshowKwargs):
+    if not incomingAxis:
+        _plt.tight_layout()
+    return _plt.gcf()
+
+def Histogram2DErrors(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlabel="", ylabel="", zlabel="",
+                      title="", aspect="auto", scalingFactor=1.0, xScalingFactor=1.0, yScalingFactor=1.0,
+                      figsize=(6,5), vmin=None, autovmin=False, vmax=None, colourbar=True, ax=None,
+                      cax=None, **imshowKwargs):
     """
     Similar to Histogram2D() but plot the errors from the histogram instead of the contents.
     See pybdsim.Plot.Histogram2D for documentation of arguments.
@@ -795,7 +825,8 @@ def Histogram2DErrors(histogram, logNorm=False, xLogScale=False, yLogScale=False
     h2 = _copy.deepcopy(histogram)
     h2.contents = h2.errors # set contents as errors and just use regular plot
     return Histogram2D(h2, logNorm, xLogScale, yLogScale, xlabel, ylabel, zlabel, title, aspect, scalingFactor,
-                       xScalingFactor, yScalingFactor, figsize, vmin, autovmin, vmax, **imshowKwargs)
+                       xScalingFactor, yScalingFactor, figsize, vmin, autovmin, vmax,
+                       colorbar, ax, cax, **imshowKwargs)
 
 def Histogram3D(th3):
     """